chore(v3): clean-code pass — dedup edge calls, event phase, CORS; update tracker
DRY pass over the V3 session code (no behavior change):
- src/lib/edgeFunction.ts: one callEdgeFunction(name, body) helper — replaces the
  duplicated session+baseUrl+fetch boilerplate in gym.ts, billing.ts and
  adminChallenges.ts
- src/features/pro/lib/eventPhase.ts: shared eventPhase() — replaces the phaseOf()
  copy in EventsList + EventDetail
- supabase/functions/_shared/cors.ts: shared CORS headers — removes the per-function
  cors.ts copies in create-gym + stripe-checkout
- docs/issues/issues.md: V3 marked done (all issues closed, migrations 028–043),
  V4/V5 epics added

Verified: typecheck / lint / 222 tests / build green; create-gym (201) and
stripe-checkout (200) re-deployed and smoked with the shared CORS.
