+++ src/features/pro/services/gym.ts
@@ -1,4 +1,4 @@
-import { supabase } from '@/lib/supabase';
+import { callEdgeFunction } from '@/lib/edgeFunction';
 
 /** A gym just created via KYB verification. */
 export type CreatedGym = {
@@ -16,32 +16,7 @@ export type CreatedGym = {
  * caller to gym_admin. Throws with the server `code` (e.g. `company_not_found`) on failure.
  */
 export async function createGym(input: { siret: string }): Promise<CreatedGym> {
-  const {
-    data: { session },
-  } = await supabase.auth.getSession();
-  if (!session?.access_token) {
-    throw new Error('no_session');
-  }
-
-  const baseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL?.replace(/\/$/, '') ?? '';
-  const anonKey = process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY ?? '';
-  if (!baseUrl || !anonKey) {
-    throw new Error('server_misconfigured');
-  }
-
-  const res = await fetch(`${baseUrl}/functions/v1/create-gym`, {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-      Authorization: `Bearer ${session.access_token}`,
-      apikey: anonKey,
-    },
-    body: JSON.stringify({ siret: input.siret }),
-  });
-
-  const parsed = (await res.json().catch(() => ({}))) as { code?: string; gym?: CreatedGym };
-  if (!res.ok || !parsed.gym) {
-    throw new Error(parsed.code ?? `http_${res.status}`);
-  }
+  const parsed = await callEdgeFunction<{ gym?: CreatedGym }>('create-gym', { siret: input.siret });
+  if (!parsed.gym) throw new Error('gym_create_failed');
   return parsed.gym;
 }
