+++ supabase/functions/create-gym/index.ts
@@ -6,7 +6,7 @@
 
 import { createClient } from 'npm:@supabase/supabase-js@2';
 
-import { corsHeaders } from './cors.ts';
+import { corsHeaders } from '../_shared/cors.ts';
 
 const REGISTRY_URL = 'https://recherche-entreprises.api.gouv.fr/search';
